requirements.txt
flask
opencv-python
ultralytics
numpy
easyocr
google-generativeai
streamlit
plotly
Pillow
python-dotenv